Only assets, all assets in assets/
Not JS source files that aren't loaded by any pages but are instead
built into bundle.js.

Also should contain materialize.css (which we symlink to
node_modules/<whatever>/materialize.css rather than vendoring so that
it can still be updated as easily as `npm update --save`).

Note that this now means the exact files to run the grapher are just:
- index.html
- bundle.js
- assets/**